Improve multithreaded building, few changes for BUILD_ALL and removing debug code from MOTD

+++ lib/image-helpers.sh
@@ -125,11 +125,9 @@ install_deb_chroot()
 	local package=$1
 	local name
 	name=$(basename "${package}")
-	cp "${package}" "${SDCARD}/root/${name}"
+	[[ ! -f "${SDCARD}/root/${name}" ]] && cp "${package}" "${SDCARD}/root/${name}"
 	display_alert "Installing" "$name"
 	[[ $NO_APT_CACHER != yes ]] && local apt_extra="-o Acquire::http::Proxy=\"http://${APT_PROXY_ADDR:-localhost:3142}\" -o Acquire::http::Proxy::localhost=\"DIRECT\""
 	LC_ALL=C LANG=C chroot "${SDCARD}" /bin/bash -c "DEBIAN_FRONTEND=noninteractive apt -yqq \
 		$apt_extra --no-install-recommends install ./root/$name" >> "${DEST}"/debug/install.log 2>&1
-
-	rm -f "${SDCARD}/root/${name}"
 }
